Refactoring folder organization
